+++ assemblyadapter-list/src/main/java/com/github/panpf/assemblyadapter/list/AssemblyListAdapter.kt
@@ -15,12 +15,14 @@
  */
 package com.github.panpf.assemblyadapter.list
 
+import android.database.DataSetObserver
 import android.view.View
 import android.view.ViewGroup
 import android.widget.BaseAdapter
 import com.github.panpf.assemblyadapter.*
 import com.github.panpf.assemblyadapter.internal.ItemDataStorage
 import com.github.panpf.assemblyadapter.internal.ItemFactoryStorage
+import com.github.panpf.assemblyadapter.list.internal.AdapterDataObservable
 
 /**
  * An implementation of [BaseAdapter], which implements multi-type adapters through standardized [ItemFactory].
@@ -35,11 +37,12 @@ import com.github.panpf.assemblyadapter.internal.ItemFactoryStorage
 open class AssemblyListAdapter<DATA>(
     itemFactoryList: List<ItemFactory<*>>,
     initDataList: List<DATA>? = null,
-    private val hasStableIds: Boolean = false,
 ) : BaseAdapter(), AssemblyAdapter<ItemFactory<*>> {
 
     private val itemFactoryStorage = ItemFactoryStorage(itemFactoryList)
     private val itemDataStorage = ItemDataStorage(initDataList) { notifyDataSetChanged() }
+    private var hasStableIds = false
+    private val adapterDataObservable = AdapterDataObservable()
 
     /**
      * Get the current list. If a null list is submitted through [submitList], or no list is submitted, an empty list will be returned.
@@ -67,12 +70,30 @@ open class AssemblyListAdapter<DATA>(
         return itemDataStorage.getData(position)
     }
 
+    /**
+     * Indicates whether each item in the data set can be represented with a unique identifier
+     * of type [java.lang.Long].
+     *
+     * @param hasStableIds Whether items in data set have unique identifiers or not.
+     * @see hasStableIds
+     * @see getItemId
+     */
+    fun setHasStableIds(hasStableIds: Boolean) {
+        if (hasObservers()) {
+            throw IllegalStateException(
+                "Cannot change whether this adapter has "
+                        + "stable IDs while the adapter has registered observers."
+            )
+        }
+        this.hasStableIds = hasStableIds
+    }
+
     override fun hasStableIds(): Boolean {
         return hasStableIds
     }
 
     override fun getItemId(position: Int): Long {
-        return if (hasStableIds) {
+        return if (hasStableIds()) {
             val data = getItem(position) ?: Placeholder
             if (data is ItemId) data.itemId else data.hashCode().toLong()
         } else {
@@ -118,5 +139,24 @@ open class AssemblyListAdapter<DATA>(
         return itemFactoryStorage.getItemFactoryByData(
             data, "ItemFactory", "AssemblyListAdapter", "itemFactoryList"
         )
+    }
+
+    override fun registerDataSetObserver(observer: DataSetObserver?) {
+        super.registerDataSetObserver(observer)
+        adapterDataObservable.registerObserver(observer)
+    }
+
+    override fun unregisterDataSetObserver(observer: DataSetObserver?) {
+        super.unregisterDataSetObserver(observer)
+        adapterDataObservable.unregisterObserver(observer)
+    }
+
+    /**
+     * Returns true if one or more observers are attached to this adapter.
+     *
+     * @return true if this adapter has observers
+     */
+    fun hasObservers(): Boolean {
+        return adapterDataObservable.hasObservers()
     }
 }