+++ assemblyadapter-list/src/main/java/com/github/panpf/assemblyadapter/list/expandable/AssemblySingleDataExpandableListAdapter.kt
@@ -15,6 +15,7 @@
  */
 package com.github.panpf.assemblyadapter.list.expandable
 
+import android.database.DataSetObserver
 import android.view.View
 import android.view.ViewGroup
 import android.widget.BaseExpandableListAdapter
@@ -22,6 +23,7 @@ import com.github.panpf.assemblyadapter.*
 import com.github.panpf.assemblyadapter.internal.ItemFactoryStorage
 import com.github.panpf.assemblyadapter.ItemId
 import com.github.panpf.assemblyadapter.list.R
+import com.github.panpf.assemblyadapter.list.internal.AdapterDataObservable
 
 /**
  * Single data version of [AssemblyExpandableListAdapter]
@@ -36,10 +38,11 @@ import com.github.panpf.assemblyadapter.list.R
 open class AssemblySingleDataExpandableListAdapter<GROUP_DATA : Any, CHILD_DATA>(
     itemFactoryList: List<ItemFactory<*>>,
     initData: GROUP_DATA? = null,
-    private val hasStableIds: Boolean = false,
 ) : BaseExpandableListAdapter(), AssemblyAdapter<ItemFactory<*>> {
 
     private val itemFactoryStorage = ItemFactoryStorage(itemFactoryList)
+    private var hasStableIds = false
+    private val adapterDataObservable = AdapterDataObservable()
 
     var data: GROUP_DATA? = initData
         set(value) {
@@ -52,11 +55,29 @@ open class AssemblySingleDataExpandableListAdapter<GROUP_DATA : Any, CHILD_DATA>
     constructor(
         itemFactory: ItemFactory<GROUP_DATA>,
         initData: GROUP_DATA? = null,
-        hasStableIds: Boolean = false,
-    ) : this(listOf(itemFactory), initData, hasStableIds)
+    ) : this(listOf(itemFactory), initData)
 
     init {
         require(itemFactoryList.isNotEmpty()) { "itemFactoryList Can not be empty" }
+    }
+
+    /**
+     * Indicates whether each item in the data set can be represented with a unique identifier
+     * of type [java.lang.Long].
+     *
+     * @param hasStableIds Whether items in data set have unique identifiers or not.
+     * @see hasStableIds
+     * @see getGroupId
+     * @see getChildId
+     */
+    fun setHasStableIds(hasStableIds: Boolean) {
+        if (hasObservers()) {
+            throw IllegalStateException(
+                "Cannot change whether this adapter has "
+                        + "stable IDs while the adapter has registered observers."
+            )
+        }
+        this.hasStableIds = hasStableIds
     }
 
     override fun hasStableIds(): Boolean = hasStableIds
@@ -73,7 +94,7 @@ open class AssemblySingleDataExpandableListAdapter<GROUP_DATA : Any, CHILD_DATA>
     }
 
     override fun getGroupId(groupPosition: Int): Long {
-        return if (hasStableIds) {
+        return if (hasStableIds()) {
             val data = getGroup(groupPosition)
             if (data is ItemId) data.itemId else data.hashCode().toLong()
         } else {
@@ -260,5 +281,24 @@ open class AssemblySingleDataExpandableListAdapter<GROUP_DATA : Any, CHILD_DATA>
         return itemFactoryStorage.getItemFactoryByData(
             data, "ItemFactory", "AssemblyExpandableListAdapter", "itemFactoryList"
         )
+    }
+
+    override fun registerDataSetObserver(observer: DataSetObserver?) {
+        super.registerDataSetObserver(observer)
+        adapterDataObservable.registerObserver(observer)
+    }
+
+    override fun unregisterDataSetObserver(observer: DataSetObserver?) {
+        super.unregisterDataSetObserver(observer)
+        adapterDataObservable.unregisterObserver(observer)
+    }
+
+    /**
+     * Returns true if one or more observers are attached to this adapter.
+     *
+     * @return true if this adapter has observers
+     */
+    fun hasObservers(): Boolean {
+        return adapterDataObservable.hasObservers()
     }
 }