+++ assemblyadapter-list/src/main/java/com/github/panpf/assemblyadapter/list/internal/AdapterDataObservable.kt
@@ -1,0 +1,10 @@
+package com.github.panpf.assemblyadapter.list.internal
+
+import android.database.DataSetObserver
+import android.database.Observable
+
+class AdapterDataObservable : Observable<DataSetObserver>() {
+    fun hasObservers(): Boolean {
+        return mObservers.isNotEmpty()
+    }
+}