+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 list:
 * new: ExpandableChildItemFactory added setOnChildViewClickListener, setOnChildViewLongClickListener, setOnChildItemClickListener, setOnChildItemLongClickListener method
 * change: ItemId move to common-core module
+* change: hasStableIds changed from constructor attribute to setHasStableIds() method
 
 # v4.0.0-beta02